deleted wiki pages that I don't endorse
It's not required or even recommended to coordinate on a signup sheet
before starting on the C backend. If folks want to do that, that's fine,
but this made it look like it was official.

The contributing to stage2 article had outdated information and other
recommendations that I don't endorse.